Count transaction documents for the dashboard inbound/outbound cards
overview() summed the item quantities of the four transaction tables, so
"Tong luong nhap kho" and "Tong luong xuat kho" reported kilograms and pieces
instead of the number of phieu. A single receipt of 100 kg read as 100.

Count the documents themselves: GoodsReceipt + ProductReceipt for inbound,
GoodsIssue + ProductIssue for outbound. Those are the same four repositories
the /receipts-issues endpoints page over with an unfiltered findAll, so the
cards now match the document count the Phieu nhap xuat module lists.

totalGoodsReceived/totalGoodsIssued become long, matching totalMaterials and
totalProducts. OverviewCards is the only consumer and passes both through
formatNumber, so the frontend needs no change.

+++ src/main/java/com/github/xnaut97/wms/dto/dashboard/DashboardOverviewResponse.java
@@ -15,9 +15,9 @@ public class DashboardOverviewResponse {
 
     private BigDecimal totalInventoryValue;
 
-    private BigDecimal totalGoodsReceived;
+    private long totalGoodsReceived;
 
-    private BigDecimal totalGoodsIssued;
+    private long totalGoodsIssued;
 
     private BigDecimal currentInventoryQuantity;
